feat: useProjectDetails 훅 분리
